+++ webserver/app/registries_api.py
@@ -3,6 +3,7 @@ containers endpoints:
 - GET /registries
 - GET /registries/<registry_id>
 - POST /registries
+- PATCH /registries/<registry_id>
 """
 
 from flask import Blueprint, request
@@ -54,3 +55,19 @@ def add_registry():
     registry = Registry(**body)
     registry.add()
     return {"id": registry.id}, 201
+
+
+@bp.route('/<int:registry_id>', methods=['PATCH'])
+@audit
+@auth(scope='can_admin_dataset')
+def patch_registry(registry_id:int):
+    """
+    PATCH /registries/<registry_id> endpoint.
+    """
+    registry = Registry.query.filter(Registry.id == registry_id).one_or_none()
+    if registry is None:
+        raise InvalidRequest(f"Registry {registry_id} not found")
+
+    registry.update(**request.json)
+
+    return {}, 204