+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Releases Changelog
 
+## 1.1.0
+- Added a `PATCH` endpoint for `/registries` so it's easier to update credentials
+- Added the `active` field for registries, so outdated ones can be safely deactivated
+
 ## 1.0.0
 - Added the Federated Node Task Controller as a chart dependency. This can be installed by setting `outboundMode` to true on the values file. By default, it won't be installed.
 - Some jobs will be cleaned before and after an upgrade.